Add tests for lexical reference system and improve evidence handling
- Implement tests for the lexical reference system to ensure it scores correctly based on themes and tags.
- Update the conversation handling to ensure proper type casting in SQL queries.
- Refactor decision persistence to improve clarity and functionality.
- Enhance the prompt structure to support mixed evidence formats.
- Extend validation logic to include detailed citation information and handle mixed evidence cases.
- Introduce a new route for retrieving evidence chunks and their associated PDF URLs.
- Add a modal for displaying evidence details in the frontend, including handling PDF retrieval.
- Update retrieval logic to include additional metadata and improve formatting.
- Ensure that the storage integration supports signed URLs for PDF access.

+++ server/domain/coach/conversacion.js
@@ -76,7 +76,7 @@ export async function compactarSiHaceFalta(conversationId, { db = pool, llmProvi
     `SELECT id, role, contenido FROM messages
       WHERE conversation_id = $1
       ORDER BY created_at
-      LIMIT GREATEST($2 - $3, 0);`,
+      LIMIT GREATEST($2::int - $3::int, 0);`,
     [conversationId, total, literales]
   );
   if (!antiguos.length) return { compactado: false, total };
